# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:1337",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/basestation/find": {
      "get": {
        "tags": [
          "basestation"
        ],
        "summary": "Get all Basestation",
        "description": "",
        "operationId": "getBasestation",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Documents"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Basestation not found"
          }
        }
      }
    },
    "/basestation/update": {
      "post": {
        "tags": [
          "basestation"
        ],
        "summary": "Update Basestation",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateBasestation",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update Basestation object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Basestation"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/basestation/delete": {
      "delete": {
        "tags": [
          "basestation"
        ],
        "summary": "Delete Basestation",
        "description": "",
        "operationId": "deleteBasestation",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Basestation id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
    "/connrenewal/find": {
      "get": {
        "tags": [
          "Connecion Renewal"
        ],
        "summary": "Get all Connecion Renewal",
        "description": "",
        "operationId": "getConnRenewal",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ConnecionRenewal"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "ConnRenewal not found"
          }
        }
      }
    },
    "/connrenewal/update": {
      "post": {
        "tags": [
          "Connecion Renewal"
        ],
        "summary": "Update Connecion Renewal",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateConnRenewal",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update Connecion Renewal object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConnecionRenewal"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/connrenewal/delete": {
      "delete": {
        "tags": [
          "Connecion Renewal"
        ],
        "summary": "Delete Connecion Renewal",
        "description": "",
        "operationId": "deleteConnRenewal",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Connecion Renewal id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Documents": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "file_name": {
          "type": "string"
        },
        "file_path": {
          "type": "string"
        },
        "customer_id": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Documents Status"
        }
      },
      "xml": {
        "name": "Documents"
      }
    },
    "Basestation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "lat": {
          "type": "string"
        },
        "lag": {
          "type": "string"
        },
        "bandwidth": {
          "type": "string"
        },
        "max_connection": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Basestation Status"
        }
      },
      "xml": {
        "name": "Basestation"
      }
    },
...
    "ConnecionRenewal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "renewal_date": {
          "type": "string"
        },
        "connection_id": {
          "type": "number"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Connecion Renewal Status"
        }
      },
      "xml": {
        "name": "Connecion Renewal"
      }
    }
  },
...
}